Create MCMC Functions
-Function 1 : prior(params)
prior takes a tuple of 6 or 4 floats or int representing A, B, C, D, E and F or A, B, D and E (Cf. ReadMe). It returns True if these parameters meet the requirements explicitely defined in the function. Otherwise, it returns False. It's purpose is to constrain the value of these parameters in the MCMC resolutions of the next functions.

-Function 2 : _ln_L(params, X, a, b, correction)
_ln_L takes 4 mandatory arguments. It returns the log-likelyhood of somem data X knowing the PDF defined by params. params is a tuple as defined in Function 1; X is a catalog of galaxy clusters (such as t_amas_match in the file "Import"), a and b are the lower and upper bounds of the richness bin considerd with X and correction is a boolean saying if the function is considering the renormalisation correction of the log-likelyhood.

-Function 3 : MCMC(richesse_bin, redshift_bin, itt, redshift_params = True, correction = True)
MCMC takes 3 mandatory arguments. richesse_bin and redshift_bin are tuples containing 2 float or int each, representing the bounds of a bining it will perform on t_amas_match (Cf. "Import"). itt is an int representing the number of steps each of the 25 walkers will perform until the MCMC is considered finished.